Updated spacepark model and controller, added migration Updated rest client tests and readme

+++ Source/SpaceParkAPI/Models/Spaceport.cs
@@ -15,6 +15,6 @@ namespace SpaceParkAPI.Models
         [Required]
         public string Name { get; set; }
         [Required]
-        public int MaxStarshipLength { get; set; }
+        public double MaxStarshipLength { get; set; }
     }
 }
